Relocate .css styling
