+++ react/src/styling/SidebarOption.css
